% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dist_fns.R
\name{clean_ps_and_qs}
\alias{clean_ps_and_qs}
\title{Clean up ps and qs provided by user: handle missing and unsorted values}
\usage{
clean_ps_and_qs(ps, qs)
}
\arguments{
\item{ps}{vector of probability levels}

\item{qs}{vector of quantile values correponding to ps}
}
\value{
named list with entries \code{ps} and \code{qs}
}
\description{
Clean up ps and qs provided by user: handle missing and unsorted values
}
