% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_dup_run_inds}
\alias{get_dup_run_inds}
\title{Get indices of starts and ends of runs of duplicate values}
\usage{
get_dup_run_inds(dups)
}
\arguments{
\item{dups}{a boolean vector that would result from calling
\code{duplicated_tol(..., incl_first = FALSE)}}
}
\value{
named list with entries \code{starts} giving indices of the first element
in each sequence of runs of duplicate values and \code{ends} giving indices of
the last element in each sequence of runs of duplicate values.
}
\description{
Get indices of starts and ends of runs of duplicate values
}
