% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int.R
\name{grid_augment_ps_qs}
\alias{grid_augment_ps_qs}
\title{Internal function that augments ps and qs by filling in a grid of
intermediate values for qs. n_grid new points are inserted between each
pair of consecutive values in qs, and the corresponding ps are filled in
by evaluating the spline output from \code{spline_cdf_direct} at the qs grid.}
\usage{
grid_augment_ps_qs(ps, qs, tail_dist, n_grid)
}
\description{
Internal function that augments ps and qs by filling in a grid of
intermediate values for qs. n_grid new points are inserted between each
pair of consecutive values in qs, and the corresponding ps are filled in
by evaluating the spline output from \code{spline_cdf_direct} at the qs grid.
}
