% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dist_fns.R
\name{make_d_fn}
\alias{make_d_fn}
\title{Creates a function that evaluates the probability density function of an
approximation to a distribution obtained by interpolating and extrapolating
from a set of quantiles of the distribution.}
\usage{
make_d_fn(
  ps,
  qs,
  interior_method = c("hyman", "monoH.FC"),
  lower_tail_dist = "norm",
  upper_tail_dist = "norm"
)
}
\arguments{
\item{ps}{vector of probability levels}

\item{qs}{vector of quantile values correponding to ps}

\item{interior_method}{method for monotonic cubic spline fit to the quantiles
used for approximating the distribution on the interior of the quantiles.
See \code{\link[stats:splinefun]{stats::splinefun()}}. One of \code{"hyman"} or \code{"monoH.FC"}.}

\item{lower_tail_dist}{name of parametric distribution for the lower tail}

\item{upper_tail_dist}{name of parametric distribution for the upper tail}
}
\value{
a function with arguments \code{x} and \code{log} that can be used to evaluate
the approximate density function (or its \code{log}) at the points \code{x}.
}
\description{
Creates a function that evaluates the probability density function of an
approximation to a distribution obtained by interpolating and extrapolating
from a set of quantiles of the distribution.
}
