% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext.R
\name{q_ext_factory}
\alias{q_ext_factory}
\title{Extrapolate quantile function in a location-scale family matching specified
quantiles.}
\usage{
q_ext_factory(ps, qs, dist)
}
\arguments{
\item{ps}{vector of two probability levels at which the distribution's
quantiles are distinct}

\item{qs}{vector of two distinct quantile values corresponding to the
probability levels in ps}

\item{dist}{the probability distribution to use for extrapolation. This
distribution should be in a location-scale family, such as "normal" or
"Cauchy"}
}
\value{
a function with parameter \code{p} that can be used to evaluate the
quantile function of the distribution in the specified location-scale
family that has quantiles matching those in \code{ps} and \code{qs}
}
\description{
Extrapolate quantile function in a location-scale family matching specified
quantiles.
}
