% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int.R
\name{spline_cdf}
\alias{spline_cdf}
\title{Approximate density function, cdf, or quantile function on the interior of
provided quantiles by representing the distribution as a sum of a discrete
part at any duplicated \code{qs} and a continuous part. The cdf of the continuous
part is estimated using a monotonic spline that interpolates the quantiles.
To obtain a pdf, we differentiate this spline. To obtain a quantile function,
we invert the approximation to the cdf, tracking any discontinuities.}
\usage{
spline_cdf(
  ps,
  qs,
  fn_type = c("d", "p", "q"),
  lower_tail_dist,
  upper_tail_dist
)
}
\arguments{
\item{ps}{vector of probability levels}

\item{qs}{vector of quantile values correponding to ps}

\item{fn_type}{the type of function that is requested: \code{"d"} for a pdf,
\code{"p"} for a cdf, or \code{"q"} for a quantile function.}

\item{spline_method}{the type of monotonic spline to fit: "hyman" or "monH.FC"}
}
\value{
a function to evaluate the pdf, cdf, or quantile function.
}
\description{
Approximate density function, cdf, or quantile function on the interior of
provided quantiles by representing the distribution as a sum of a discrete
part at any duplicated \code{qs} and a continuous part. The cdf of the continuous
part is estimated using a monotonic spline that interpolates the quantiles.
To obtain a pdf, we differentiate this spline. To obtain a quantile function,
we invert the approximation to the cdf, tracking any discontinuities.
}
