% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int.R
\name{spline_cdf_direct}
\alias{spline_cdf_direct}
\title{Internal function that constructs a monotonic Hermite spline interpolating
ps and qs.}
\usage{
spline_cdf_direct(ps, qs, tail_dist, fn_type = c("d", "p", "q"))
}
\description{
Internal function that constructs a monotonic Hermite spline interpolating
ps and qs.
}
