% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int.R
\name{spline_qf}
\alias{spline_qf}
\title{Approximate density function, cdf, or quantile function on the interior of
provided quantiles by using a monotonic spline that interpolates the
quantiles to estimate the quantile function, inverting the quantile function
to estimate the cdf, and differentiating the cdf to estimate the pdf.}
\usage{
spline_qf(ps, qs, lower_tail_dist, upper_tail_dist, fn_type = c("d", "p", "q"))
}
\arguments{
\item{ps}{vector of probability levels. These must be unique.}

\item{qs}{vector of quantile values correponding to ps}

\item{lower_tail_dist}{name of parametric distribution for the lower tail}

\item{upper_tail_dist}{name of parametric distribution for the upper tail}

\item{fn_type}{the type of function that is requested: \code{"d"} for a pdf,
\code{"p"} for a cdf, or \code{"q"} for a quantile function.}
}
\value{
a function to evaluate the pdf, cdf, or quantile function.
}
\description{
Approximate density function, cdf, or quantile function on the interior of
provided quantiles by using a monotonic spline that interpolates the
quantiles to estimate the quantile function, inverting the quantile function
to estimate the cdf, and differentiating the cdf to estimate the pdf.
}
