% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext.R
\name{calc_loc_scale_params}
\alias{calc_loc_scale_params}
\title{Calculate location and scale parameters for a specified distribution so that
it matches two specified quantiles}
\usage{
calc_loc_scale_params(ps, qs, dist)
}
\arguments{
\item{ps}{vector of two probability levels at which the distribution's
quantiles are distinct}

\item{qs}{vector of two distinct quantile values corresponding to the
probability levels in ps}

\item{dist}{the probability distribution to use for extrapolation. This
distribution should be in a location-scale family, such as \code{"norm"}al or
\code{"cauchy"}}

\item{named}{list with entries \code{"a"}, the location parameter, and \code{"b"}, the
scale parameter}
}
\description{
Calculate location and scale parameters for a specified distribution so that
it matches two specified quantiles
}
