% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int.R
\name{spline_cdf}
\alias{spline_cdf}
\title{Approximate density function, cdf, or quantile function on the interior of
provided quantiles by representing the distribution as a sum of a discrete
part at any duplicated \code{qs} and a continuous part for which the cdf is
estimated using a monotonic Hermite spline. See details for more.}
\usage{
spline_cdf(
  ps,
  qs,
  lower_tail_dist,
  upper_tail_dist,
  fn_type = c("d", "p", "q"),
  n_grid = 20,
  tol = 1e-06
)
}
\arguments{
\item{ps}{vector of probability levels}

\item{qs}{vector of quantile values correponding to ps}

\item{lower_tail_dist}{name of parametric distribution for the lower tail}

\item{upper_tail_dist}{name of parametric distribution for the upper tail}

\item{fn_type}{the type of function that is requested: \code{"d"} for a pdf,
\code{"p"} for a cdf, or \code{"q"} for a quantile function.}

\item{n_grid}{grid size to use when augmenting the input \code{qs} to obtain a
finer grid of points along which we form a piecewise linear approximation
to the spline. \code{n_grid} evenly-spaced points are inserted between each
pair of consecutive values in \code{qs}. The default value is 20. This can
be set to \code{NULL}, in which case the piecewise linear approximation is not
used. This is not recommended if the \code{fn_type} is \code{"q"}.}

\item{tol}{numeric tolerance for identifying duplicated quantiles indicating
a discrete component of the distribution. If there is a run of values where
each consecutive pair is closer together than the tolerance, all are
labeled as duplicates even if not all values in the run are within the
tolerance.}
}
\value{
a function to evaluate the pdf, cdf, or quantile function.
}
\description{
Approximate density function, cdf, or quantile function on the interior of
provided quantiles by representing the distribution as a sum of a discrete
part at any duplicated \code{qs} and a continuous part for which the cdf is
estimated using a monotonic Hermite spline. See details for more.
}
\details{
The cdf of the continuous part of the distribution is estimated
using a monotonic degree 3 Hermite spline that interpolates the quantiles
after subtracting the discrete distribution and renormalizing. In theory,
an estimate of the quantile function could be obtained by directly inverting
this spline. However, in practice, we have observed that this can suffer from
numerical problems. Therefore, the default behavior of this function is to
evaluate the "stage 1" cdf estimate corresponding to discrete point masses
plus monotonic spline at a fine grid of points, and use the "stage 2" cdf
estimate that linearly interpolates these points with steps at any duplicated
q values. The quantile function estimate is obtained by inverting this
"stage 2" cdf estimate. When the distribution is continuous, we can obtain an
estimate of the pdf by differentiating the cdf estimate, resulting in a
discontinuous "histogram density". The size of the grid can be specified with
the \code{n_grid} argument. In settings where it is desirable to obtain a
continuous density function, the "stage 1" cdf estimate can be used by
setting \code{n_grid = NULL}.
}
